requirements.txt
scrapy
PyPDF2
pdfminer.six
beautifulsoup4
requests
python-dotenv
tldextract
